Update README and reorganize
