+++ src/main/java/io/github/samipourquoi/endtech/mixin/MixinMinecraftServer.java
@@ -1,6 +1,6 @@
 package io.github.samipourquoi.endtech.mixin;
 
-import io.github.samipourquoi.endtech.helpers.RuleSettings;
+import io.github.samipourquoi.endtech.ETAdditionsSettings;
 import net.minecraft.server.MinecraftServer;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
@@ -11,6 +11,6 @@ import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 public class MixinMinecraftServer {
     @Inject(method = "loadWorld", at = @At("HEAD"))
     private void loadConf(CallbackInfo ci) {
-        RuleSettings.loadConfigSettings((MinecraftServer) (Object) this);
+        ETAdditionsSettings.loadConfigSettings((MinecraftServer) (Object) this);
     }
 }
